fix: change the characters location_id column type

+++ database/migrations/2025_04_05_050410_add_location_to_characters_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('characters', function (Blueprint $table) {
-            $table->unsignedSmallInteger('location_id')->after('status');
+            $table->unsignedBigInteger('location_id')->after('status');
             $table->foreign('location_id')->references('id')->on('locations');
         });
     }
